Adds additional parts
