+++ Best Practices/Challenges/Refactoring/Refactoring.Challenge/Part02.cs
@@ -1,27 +1,17 @@
 ﻿namespace Refactoring.Challenge;
 
 /// <summary>
-/// This part involves refactoring a palindrome check.
+/// This part involves refactoring a method that finds the maximum value in a list.
 /// The Run method should:
 ///     * Be refactored to use a simpler solution.
 ///     * Not break the existing test.
-///     * NOTE - The reverse of this is sometimes asked in interviews, what is the fastest way to check if a string is a palindrome.
+///     * Should not use Linq
 /// </summary>
 public class Part02
 {
-    public bool Run(string text)
+    public int Run(List<int> nums)
     {
-        var length = text.Length;
-        var charactersToCheck = length / 2;
-        
-        for (var i = 0; i < charactersToCheck; i++)
-        {
-            if (text[i] != text[length - i - 1])
-            {
-                return false;
-            }
-        }
-        
-        return true;
+        int m = 0; foreach (var n in nums) { if (n > m) m = n; }
+        return m;
     }
 }
